Added new functions and user interface code
Added a function for printing the chessboard, and an identifier to recognize the pieces on the board. (color representation is WIP)
Added a constructor for ChessGame to insert Pieces unto the board.
Added a private helper method "ParseLetter" to simply translating user-input letter positions into array positions.
Added a constructor for Piece to assign "color".
Added complete implementation of CalculateAngle for the Knight type.