Big addition to functionality
Added a function CheckMovement that iterates from current to target position and checks for obstacles, returns true or false based on whether the move is legal.
Fully implemented PieceMovement. This is the function that tells the main program whether a move is legal based on angle and the outcome of CheckMovement.
Added an if statement inside the CalculateAngle overrides for the classes Rook, Bishop, and Queen, to prevent the program from attempting to divide by zero.
Added a commented array table for visualization.
Removed redundant using directive.